Refactored code - reduced duplicate code, increased efficency
