+++ lesson.py
@@ -1,139 +1,6 @@
 # Essential Imports
 import time
-from speech import Speech
-import pygame
-from gui import Gui
-from alphabet import Alphabet
 from time import sleep
-from error_logger import ErrorLogger
-
-
-class Interaction ():
-    def __init__(self):
-        error_log = ErrorLogger()
-        using_raspberry_pi = True
-
-        try:
-            # Attempting to acces the GPIO Pins Module
-            import RPi.GPIO as GPIO
-            from character import BrailleCharacter
-        except ModuleNotFoundError as e:
-            # If the system is unable to import the module it will revert to using the SDFJKL keys
-            print("KEYBOARD MODE ACTIVATED")
-            from keyboard_interface import check_keys
-            using_raspberry_pi = False
-            error_log.log_error(e)
-            dot_has_test = {
-                "F": 0,
-                "D": 0,
-                "S": 0,
-                "J": 0,
-                "K": 0,
-                "L": 0
-            }
-
-        # If the pi is being used initiate the pins
-        if using_raspberry_pi:
-            GPIO.setmode(GPIO.BCM)
-            current_char = BrailleCharacter()
-
-        # Initialising objects
-        braille_alphabet = Alphabet()
-        speech = Speech()
-
-        # Introduction to Lesson 1 and the importance of Braile...
-        # speech.say("Welcome to your first Braille lesson!")
-        # speech.say("Braille changes lives. It gives thousands of people independence, learning, literacy, and the enjoyment of reading. Braille opens doors, and gives hope and inspiration.")
-        # speech.say("""Popular games (such as Bingo and Uno) are available in braille and others can be adapted by the addition of braille labels, enabling a blind or partially sighted individual to join in with family or friends in a wide range of leisure pursuits.""")
-        # Content source: https://www.rnib.org.uk/braille-and-moon-%E2%80%93-tactile-codes-braille-past-present-and-future/modern-day-braille
-        # Last accessed: 03/03/2021
-
-        # Attempt to show GUI
-        try:
-            show_gui = True
-            graphical_user_interface = Gui()
-        except pygame.error:
-            print("No gui available")
-            show_gui = False
-            error_log.log_error("No gui available")
-
-        self.error_log = error_log
-        self.show_gui = show_gui
-        if show_gui:
-            self.graphical_user_interface = graphical_user_interface
-
-        self.speech = speech
-        self.braille_alphabet = braille_alphabet
-
-        self.using_raspberry_pi = using_raspberry_pi
-
-        if using_raspberry_pi:
-            self.current_char = current_char
-        else:
-            self.dot_has_test = dot_has_test
-            self.check_keys = check_keys
-
-
-class Translator():
-    def __init__(self, interaction_object):
-        # print(interasction_object.speech)
-        # dot_hash = "101010"
-        # interaction_module.graphical_user_interface.draw_dot_hash(
-        #                     dot_hash, "")
-        # interaction_module.speech.say(dot_hash)
-        # dot_hash = "000000"
-        # interaction_module.graphical_user_interface.draw_dot_hash(
-        #                     dot_hash, "")
-        # interaction_module.speech.say(dot_hash)
-        # dot_hash = "111111"
-        # interaction_module.graphical_user_interface.draw_dot_hash(
-        #                     dot_hash, "")
-        # interaction_module.speech.say(dot_hash)
-        # dot_hash = "000000"
-        # interaction_module.graphical_user_interface.draw_dot_hash(
-        #                     dot_hash, "")
-        # interaction_module.speech.say(dot_hash)
-
-        previous_time = time.time()
-        start_time = time.time()
-
-        # Forever loop which will ONLY exit if the program is ended
-        expired_events = []
-        previous_letter = "_"
-
-        # if the lesson is still live then this loop will continue
-        lesson_live = True
-        while lesson_live:
-            # Check if enough time has passed to update GUI
-            now = time.time()
-            difference = float(now-previous_time)
-
-            # Checks timelines every 0.5 seconds
-            if difference > 0.5:
-                print(difference)
-                # speech.say(current_dots_hash)
-                previous_time = now
-
-                time_since_start = float(now - start_time)
-                print(difference, time_since_start)
-                if interaction_module.using_raspberry_pi:
-                    # Get dot hash from pins
-                    current_dots_hash = (
-                        interaction_module.current_char.get_current_dots_hash())
-                else:
-                    # Get dot hash from keyboard
-                    current_dots_hash = interaction_module.check_keys(
-                        pygame, interaction_module.dot_has_test)
-                braille_translation = interaction_module.braille_alphabet.translate_braille_to_alphabet(
-                    current_dots_hash)
-                previous_time = now
-                if interaction_module.show_gui:
-                    interaction_module.graphical_user_interface.draw_dot_hash(
-                        current_dots_hash, braille_translation)
-
-                if braille_translation != previous_letter:
-                    previous_letter = braille_translation
-                    interaction_module.speech.say(braille_translation)
 
 
 class Lesson():
@@ -278,6 +145,7 @@ class Lesson():
         self.speech = interaction_object.speech
         self.using_raspberry_pi = interaction_object.using_raspberry_pi
         self.braille_alphabet = interaction_object.braille_alphabet
+        self.pygame = interaction_object.pygame
 
         if not interaction_object.using_raspberry_pi:
             self.current_char = ""
@@ -339,7 +207,7 @@ class Lesson():
                 else:
                     # Get dot hash from keyboard
                     current_dots_hash = self.check_keys(
-                        pygame, self.dot_has_test)
+                        self.pygame, self.dot_has_test)
                 braille_translation = self.braille_alphabet.translate_braille_to_alphabet(
                     current_dots_hash)
                 previous_time = now
@@ -351,8 +219,3 @@ class Lesson():
 
         self.graphical_user_interface.draw_dot_hash(
             current_dots_hash, "")
-
-
-interaction_module = Interaction()
-lesson_1 = Lesson(interaction_module)
-translator = Translator(interaction_module)
